refactor(plugins): flatten builtin template folders
